bug 389 fixed
git-svn-id: http://co19.googlecode.com/svn/trunk/co19/tests/co19@681 2ecb45c1-dfb3-dc95-330a-13f216d23a04

+++ src/LibTest/core/double/ceil_A01_t01.dart
@@ -4,27 +4,26 @@
  * BSD-style license that can be found in the LICENSE file.
  */
 /**
- * @assertion double ceil()
+ * @assertion abstract int ceil()
+ * Returns the least integer no smaller than this.
  * @description Checks that positive values less then one are rounded to one.
  * @author pagolubev
  * @reviewer msyabro
  */
+import "dart:math" as Math;
 import "../../../Utils/expect.dart";
 
-
-import "dart:math" as Math;
-
-check(double expected, double argument) {
+check(int expected, double argument) {
   Expect.equals(expected, argument.ceil());
 }
 
 main() {
-  check(1.0, .1);
-  check(1.0, .01);
-  check(1.0, 1.0E-100);
-  check(1.0, 4.9406564584124654e-324); // min value
-  check(1.0, 2.2250738585072014e-308); // min normal
-  check(1.0, .5);
-  check(1.0, .9);
-  check(1.0, 1.0 - Math.pow(2.0, -53));
+  check(1, .1);
+  check(1, .01);
+  check(1, 1.0E-100);
+  check(1, 4.9406564584124654e-324); // min value
+  check(1, 2.2250738585072014e-308); // min normal
+  check(1, .5);
+  check(1, .9);
+  check(1, 1.0 - Math.pow(2.0, -53));
 }
